requirements.txt
django==1.8
djangorestframework
